[IMP] stock_barcode: digipad to edit line qty.
In the barcode line edit form view, adds a digipad to write quantity
faster. This digipad isn't displayed in case the line is about a product
tracked by serial (as the max qty. for a product tracked by SN is 1).

This `digipad` is a new widget and has a property named `quantity_field`
who must be the name of the numeric field the widget will have access to

Also, if the product's line has product packagings, the digipad will add
a button for the first three packagings. These buttons can be used to
increment the quantity by the packaging's quantity.

task-2394784

Part-of: odoo/enterprise#15755