[FIX] web_studio: avoid traceback on auto suggest studio field
Before this commit: if user adds integer field with name x_studio_partner_id
and then tries to edit the document and save it theows traceback, it is because
we handle 'x_studio_partner_id' field specifically and if it is available in
self._fields then we assumed that it will always have relation with res.partner
so if integer field is added with 'x_studio_partner_id' name then it will throw
tracback because we consider it as many2one.

After this commit: if user adds integer with 'x_studio_partner_id' field then
consider it as integer field and do not call _message_add_suggested_recipient
on integer field, so to fix this we added check that 'x_studio_partner_id'
field is type of many2one and it's comodel_name is 'res.partner' to make sure
that it is relational many2one field with res.partner.

task-2538257

closes odoo/enterprise#20378

X-original-commit: 7ca0cc9750cece08eb0eaf12976a6e21393d62b5
Signed-off-by: Nicolas Lempereur (nle) <nle@odoo.com>