[FIX] helpdesk: fix kanban state label according to legends on stage
Currently, when open the kanban view of the ticket and do the
quick create it will not display the kanban state label according
to the legends of the stage.

so in this commit, add the stage_id and legend field in quick create
view as invisible so it will get the value in state_selection widget
to display on kanban_state dropdown.

PS: legend fields are related field of stage_id so added the stage_id
and team_id is used as domain on stage_id field so added team_id too.

closes odoo/enterprise#20336

Taskid: 2531595
X-original-commit: 38d8b8bff6d11857692c8099b874f7d7d211b30c
Signed-off-by: Yannick Tivisse (yti) <yti@odoo.com>