[FIX] l10n_de_pos_cert: add invoiced orders to Fiskaly
Before this commit, the invoiced orders were not considered when validating the session with Fiskaly.

Without l10n_de_pos_cert, the normal behaviour at the validation of a session is to change the state of 'paid' orders to 'done'. However, 'invoiced' orders do not change their states. These orders also need to be taken into account.

opw-2619365

closes odoo/enterprise#20283

X-original-commit: ca9f3fbe71fcb4d20c64fe1e850c679e48365c9d
Signed-off-by: pimodoo <pimodoo@users.noreply.github.com>