[IMP] l10n_mx_reports: use account codes instead of tags_is
Replace the tags_id used in domains in the financial reports by the account ids.
Tags ids are not required and could be missing on accounts, using account ids
helps to have financial reports staying as correct as possible.

Also add control domains on the report to helps to detect misconfigurations.

Task id #2612900

closes odoo/enterprise#23383

Related: odoo/odoo#82532
Signed-off-by: Quentin De Paoli <qdp@odoo.com>