[FIX] l10n_de_pos_cert: empty payment_types Fiskaly
When every order in a session has a total of 0.0$, the 'payment_types' field is empty. This leads to a Traceback error when validating the pos session.

Before this commit, when all orders in a session are of 0.0$, we sent an empty list in the 'payment_types' field to Fiskaly, resulting in a traceback when the user needs to validate his session:
'requests.exceptions.HTTPError: 400 Client Error: Bad Request for url: https://dsfinvk.fiskaly.com/api/v0/cash_point_closings/XXXXXXXX-XXXX-XXXX-XXXX-XXXXXXXXXXXX'
Here is the complete response from Fiskaly:
'{"code":"E_FAILED_SCHEMA_VALIDATION","message":"Payment types are empty","status_code":400,"error":"Bad Request"}'

opw-2620199

closes odoo/enterprise#20263

X-original-commit: 74af885722f7edb99fa6b5d05c4100ec20ee208b
Signed-off-by: pimodoo <pimodoo@users.noreply.github.com>