[IMP] sale_timesheet_enterprise: filter timesheets in task based on config
Before this commit, the portal user can see all timesheets created in
the task. But we cannot check if he can see all timesheets or only the
ones validated. Indeed, in the configuration of Timesheet app, if
'invoice policy' is set to 'Approved timesheets only' then the portal
user could only the validated timesheets.

This commit overrides the read method in project.task, to check it and
filter if the ‘invoicing policy’ is configured accordingly in
Timesheets app.

task-2379518

Community PR: odoo/odoo#73341
Enterprise PR: odoo/upgrade#2621

closes #19483