[IMP] stock_barcode: nomenclature field in setting
Before this commit, in the Inventory settings the Barcode Nomenclature
field was shown only if the user is in debug mode.
With this commit, this field will be shown once there is at least two
barcode nomenclatures.
As `barcodes_gs1_nomenclature` is installed with `stock_barcode`, it
should have always at least two existing nomenclatures but as the user
can delete one, the condition still can be relevant.

Also, the "open and edit" button next to the nomenclature field is
removed because this form view is technical. Before this commit, as the
field was shown only when the user is in debug mode, that was OK, but
now, as the field is also shown when not in debug mode and as we can't
hide dynamicaly the open button, it's better to always hide it.
Anyway, the barcode's nomenclatures are still editable through the
configuration menu (Inventory > Configuration > Barcode Nomenclatures).

closes odoo/enterprise#19226

Related: odoo/odoo#72685
Related: odoo/upgrade#2713
Signed-off-by: Arnold Moyaux <amoyaux@users.noreply.github.com>