[FIX] worksheet,*: cannot read ir.model
* = industry_fsm_report and quality_check_worksheet

Before this commit, the access read to ir.model has been removed for
users. So, the project user could not access to worksheet and
cannot select a worksheet for fsm task.

This commit adds `sudo` when it is necessary and adds new field to get
the model in ir.model to use it in ir.rule defined in quality_control
and industry_fsm modules.

Related PR: odoo/enterprise#17633

Upgrade PR: odoo/upgrade#2744

closes #20251

Signed-off-by: Yannick Tivisse (yti) <yti@odoo.com>