[FIX] website_calendar: fix appointment slots tree view UI glitch
This commit fixes a minor UI glitch within the editable tree view of
calendar.appointment.slots.

Indeed, the current implementation uses custom classes and tags to try to alter
the size of the tree view columns, which results in strange UI display.

Trying to alter the tree view column sizes is generally not a good idea and not
clearly supported.

This commit removes the various customizations to fall back to a more standard
tree view definition.

Task-2625696

closes odoo/enterprise#20239

Signed-off-by: Thibault Delavallee (tde) <tde@openerp.com>