[IMP] helpdesk: improve the random/balance assignation of tickets.
Purpose of the commit is to improve the random/balance assignation
methods of ticket based on working calendar and time off of the member
to handle the ticket in a timely manner.

so in this commit, assign the ticket based on their resource calendar
and on their time off.
If none of the team members are available, fallback on the following period
until the next 7 days.

task-2515170

closes odoo/enterprise#20940

Signed-off-by: Laurent Stukkens (ltu) <ltu@odoo.com>