[REV] l10n_be_hr_contract_salary: Prevent choosing an invalid car
This reverts commit be8697c3d531decadbc4175fe5d0f8e3f7bb4647.

The commit above does not actually solve the issue it was created for,
instead it removes a feature we use to allow to force a car when signing
new contracts. Reverting this commit while we find a better way to solve
the issue, which is actually linked to not being able to sign with a
forced car when no user has been created yet.

Task ID: 2624963

closes odoo/enterprise#20364

X-original-commit: daf2b66383454e2d01756cbeb897622e2d430ace
Signed-off-by: Kevin Baptiste <kba@odoo.com>