[FIX] account_followup: remove reversed invoices from unpaid invoices
In follow-up reports, reversed invoices pdf's can be in attachment of follow up mails. It is confusing
because these invoices do not show in the mail but their pdf's is attached.

Before, these attachments were based on the "unpaid_invoices" of the partner which include reversed invoices.

Now, this fix will not include invoices with a "payment_state" as 'reversed' when computing "unpaid_invoices".

opw-2606629

closes odoo/enterprise#21642

X-original-commit: 1999faf35efc64c655a0918d7825fbe346414025
Signed-off-by: Laurent Smet <smetl@users.noreply.github.com>