[FIX] timesheet_grid: Fix status display with long task names
The status colored dot is squished when the name of the task is too long.
Steps to reproduce:
https://nimb.ws/aJQF47
go to timesheets > all timesheets
create a timesheet entry with a long task name and view this entry in the kanban view

The status colored dot has been placed between <div></div> tags to avoid being squished.
The issue has only been identified on the master branch.

Task-2626887

closes odoo/enterprise#20280

Signed-off-by: Yannick Tivisse (yti) <yti@odoo.com>