[IMP] {account,l10n_{be,lu}}_reports: add accounting firm to company data and xml exports
A company may grant a mandate to an accounting firm or an agent
to file its tax return or financial reports.

The goal is to allow companies to add their agent as a partner
and select it in the company data configuration.

While exporting the XML file, if an agent has been configured,
it will be used by default as the company's representative.

Task 2526717