[FIX] l10n_be_hr_payroll_fleet: assign correct company_id to contract
With the addition of multi-company rules (https://github.com/odoo/odoo/pull/73654/commits/fca3128babb0910a72d1cacda193c064df58337f),
an error was raised during the hr_contract_salary_tour.
This is because the vehicle contract was not configured with the same
company_id as the vehicle it was linked to. Whenever a user is trying to
access the vehicle infos, if he doesn't have access to both of the
companies, a multi-company error is raised.

This change should assign the vehicle's company_id to the linked contract and
avoid that error.

opw-2518188

closes odoo/enterprise#20317

X-original-commit: 560a8ef2fb6168a638b4083f0af872613afd8eb5
Signed-off-by: Yannick Tivisse (yti) <yti@odoo.com>