[IMP] industry_fsm_report: cannot count worksheet in project sharing
Before this commit, if the user has no access rights to worksheet set in
the task. Then he cannot access to the task form view because he has not
the access rights to access to this worksheet.

This commit adds sudo if the user is a portal one in the
`_compute_worksheet_count` method to continue the calculation and allows
the portal user to access to task form view.

task-2379518

Community PR: odoo/odoo#73341
Upgrade PR: odoo/upgrade#2621

closes #19483