[IMP] l10n_be_reports: add comptes de régularisation in belgian BS
The goal is to adjust the "Comptes de régularisation" section
of the belgian balance sheet.

The 496000 account "Foreign currency translation differences - Assets"
was not included.

At the same time, every control domain that includes an 'account_id.code'
was removed.

Only the ones using an 'account_id.user_type_id.internal_group'
are meant to be used from now on.

closes odoo/enterprise#20351

Signed-off-by: William André (wan) <wan@odoo.com>