[IMP] web_enterprise,web_mobile: implement partner's vCard
In the process of replacing the native methods exposed in the mobile
apps, this commit implements the download of a vCard containing a
partner's informations.

By using this standard format, both regular web users and mobile ones
are now able to save the partner's details to use them with their usual
address book software.

On a mobile device, the actual import of those informations is delegated
to the operating system.

References:
- https://datatracker.ietf.org/doc/html/rfc6350
- https://en.wikipedia.org/wiki/VCard
- https://github.com/eventable/vobject#vcards

Task ID: 2583916

closes odoo/enterprise#20461

Related: odoo/upgrade#2795
Signed-off-by: Pierre Paridans <pparidans@users.noreply.github.com>