[REF] documents_spreadsheet: Convert actions to wowl action
This commit converts the spreadsheet and template client actions
to new wowl actions.

Note: there are two "service-like" classes introduced in this commit.
They act as an interface between the actions and the server to hide
model specificities ("documents.document" vs "spreadsheet.template")
There are not actually registered as services since they are used only once.
Better naming ideas for those classes are welcome.

TODO (in the future)
- convert helpers to use the new orm service
- remove action inheritance (abstract spreadsheet action).
  There are better ways to achieve the same thing without inheritance
  which might not be straight forward to read and a bit fragile.

Task 2621549

Part-of: odoo/enterprise#20375
Co-authored-by: fleodoo <fle@odoo.com>