[FIX] web_mobile: disable native DatePicker on iOS
On the mobile iOS app, the native datetime picker is used to select the
value of a date{time} field.

Sadly, a bug in the positioning of the native picker can result in
placing it too low on the device's screen and hiding the time part of
the picker (which makes it unreachable for the user).

This issue was fixed in odoo/mobile-ios@8266a7be9529406b0f8dc5aae21b3ee6249e99b7

Due to the current situation with Apple, as we are not able to release
new version of the mobile iOS app, we don't have much choice than
disabling the native datetime picker for this platform and fallback to
the regular web-based one.

Note: only 'datetime' pickers should be affected ('date' ones works as
before).

Task ID: 2613810

closes odoo/enterprise#20446

X-original-commit: 5cec2a44d0c20af6db780ea01fde3c60725d8166
Related: odoo/odoo#75577
Signed-off-by: Adrien Dieudonné (adr) <adr@odoo.com>
Signed-off-by: Pierre Paridans <pparidans@users.noreply.github.com>