[FIX] pos_l10n_se: fix data sent to blackbox
It was impossible to send an order with the sweden blackbox.

We figured out that we sent the vat number instead of the registry
company.

We corrected the data sent from the js to the driver with the registry
company number instead of the vat number. We also added a regex to
remove the "-" from the number in the standard format.
 ex: 556677-1234 -> 5566771234

closes odoo/enterprise#20388

X-original-commit: 65de6570992f71ff2fb902e2ae1b471995428832
Signed-off-by: pimodoo <pimodoo@users.noreply.github.com>