[FIX] helpdesk: correct calculation of happy rating
Since commit https://github.com/odoo/enterprise/commit/e987fe1b9aa7f09815acd2d15ed99f4bd64fe7a8, we changed the metrics to compute the happy rating.
We should also adapt the corresponding formulas so we keep a correct calculation scale.

Now, a "great" rating is scored as 5 instead of 10. If you have only 1 rating, you should
anyway expect 100% happy rating.

The calculation using the old scoring was :
( 10 / 1 ) * 10 = 100%

With the new scoring, before this fix, we are getting :
( 5 / 1 ) * 5 = 25%

After the fix, we will get instead :
( 5 / 1 ) * 20 = 100%

If we divide the scoring by 2, we need to multiply the result in the formula instead of dividing it.

closes odoo/enterprise#20436

X-original-commit: 9e3f06f6955f201ed454043fe1c8cb40621f42a6
Signed-off-by: Alex Tuyls <alt-odoo@users.noreply.github.com>