[FIX] timesheet_grid: fix multi-company issue in timesheet
This commit is a forward port of #21139

This commit fixes a multi-company issue with the timesheet timer widget.
The domain filtering the selectable projects within the timesheet grid
timer widget has been extended to only include projects in companies
that are using hour encoding. Also prevents using the button at the
start of the grid to start the timer on project which do not encode
timesheets in hours.

How to reproduces:
- Have one company with timesheet encoding in hours, and one in days
- Create a project in the company with days encoding
- Make the hour encoding company your main company
- Activate the second company, so both are active
- Start a timer and select the project created in the days encoding
  company

Task ID: 2563519

closes odoo/enterprise#21189

X-original-commit: 5f8ed5fbb2f2e5bd7da0f8ea4c095ec1a4c35479
Signed-off-by: Arnaud Joset <arj-odoo@users.noreply.github.com>
Signed-off-by: LTU-Odoo <IT-Ideas@users.noreply.github.com>