[IMP] stock_barcode: scan package type
Package type can now be scanned in the Barcode Application.
When a barcode type is scanned alone:

 - If the selected barcode line has already a package, it will change
   this package's type for the scanned one.

 - If the selected barcode line has no package, it will do a put in pack
   in a new package of this type.

Also, the user can now scan a barcode beginning with a package type's
barcode to put in pack in a new package of this type with the scanned
barcode as package's name.

For example, if you have a package type "Palette" with "PAL" as barcode
and scan the barcode "PAL0000178", if this barcode doesn't match with
something already existing, it will check if the barcode begins with a
package type's barcode (here the palette) and then will put in pack the
scanned products in a package with the barcode as name and with the
corresponding package type.

task-2627002

Part-of: odoo/enterprise#20482