[REF] documents_spreadsheet: Rename active_id action param
The spreadsheet client action has a param `active_id` which is the
record's id.
`active_id` has a special meaning in odoo views (meaning that is kind of lost
btw).

To avoid any confusion, the param is renamed to `spreadsheet_id`.

Backward compatibility with the deprecated param is kept if a user has an
hardcoded URL somewhere with this param.

closes odoo/enterprise#20375

Signed-off-by: Pierre Rousseau (pro) <pro@odoo.com>