[FIX] worksheet: open studio on Design Template button
Since owl web framework is merged, part of studio is converted to new owl
framework design and hence, previously event were bound on widget itself and
now events are bound on DOM element, same way 'studio-icon-clicked' event is
also bound on window.

The issue here is, when we click on 'Design Template' button in Worksheet Templates
form, previous form element is destroyed and new element is pushed into DOM,
so button which is clicked is removed from DOM and it's parent form view element
is also removed, so trigger_up will not propagate 'studio-icon-clicked' event
to window, due to which clicking on 'Design Worksheet Template' do not open
studio.

With this commit: we explicitly dispatching 'studio-icon-clicked' event on window,
so when 'Design Worksheet Template' is clicked and new form is loaded, studio is
opened automatically.

task-2593520

closes odoo/enterprise#22153

X-original-commit: c8c79d6704cf49b9c67b361a4b05a9ea76ff9dad
Signed-off-by: Aaron Bohy (aab) <aab@odoo.com>