[FIX] account_bank_statement_import_camt: correct IBAN filtering
In case a statements file is targetting multiple journals, we extract only those
related to the current journal bank account. However we should sanitize the account
number to properly filter them and allow a correct import.

The corresponding test was also amended as the IBAN account number data should be
expected in a sanitized format : [a-zA-Z]{2,2}[0-9]{2,2}[a-zA-Z0-9]{1,30}.

To manage both cases, we should sanitize the account number from the XML anyway.

X-original-commit: b9450e584f343312856b96e145bdee927c5b6dad
Part-of: odoo/enterprise#21263