[IMP] sale_planning: generate shift on SO confirmation
modules: sale_planning,project_timesheet_forecast_sale

With this commit, shifts are generated once the SO is confirmed.

The shift are unassigned and doesn't have a start nor an end datetime
(it's called an unscheduled shift).

The shifts are only generated once. So if a user decides to remove it,
it won't be regenerated automatically.

The allocated hours on this unscheduled shift always follows the number
of hours to plan on the SOL minus the number of hours already planned.

The number of hours already planned on a SOL take into account the
validated timesheets and the allocated hours on the slots planned after
the last validated timesheet on this SOL.

Users will be able to plan it from the gantt view or automatically
assign it, see next commits for further details.

PR community : odoo/odoo#75553
PR enterprise : #19647
PR upgrade : odoo/upgrade#2777

task-2559061 - part:3