[FIX] l10n_cl_edi: Handle unexpected SII response
_analyze_sii_result interprets the response from the SII and does this
by checking a bunch of conditions and known response codes. If a code
occurs that it doesn't recognize (like 107 instead of 106) or if the XML
contents are different than expected, it will implicitly return None as
the function reaches its end without explicitly returning anything.

This causes problems because it will result in the SII status getting
cleared and the invoice getting stuck in that state.

This fix prevents that from happening by aborting the operation if we
don't recognize the response. This means the SII status will remain the
same as it was before. This also assumes that the unexpected response
is temporary and only happens intermittently. Another attempt will occur
the next time the related cron job is scheduled to run. If the
unexpected response turns out to be more regular, handling it should be
implemented properly in _analyze_sii_result.

closes odoo/enterprise#20898

Ticket: 2524452
X-original-commit: 21a3d84e3efeb9b0257cf32ec5057725ff5ffd4a
Signed-off-by: Josse Colpaert <jco@openerp.com>
Signed-off-by: Raf Geens <raf-odoo@users.noreply.github.com>