[REF] documents_spreadsheet: Add generic standalone many2one field
The current `ModelSelectorWidget` is actually a many2one field
to select "ir.model" records. With task 2572184, we need the same
widget to select "ir.menu" records.

This commit extract the common logic from `ModelSelectorWidget`
to a new `StandaloneMany2OneField` widget that can be used with
an arbitrary model.

Task 2572184

Part-of: odoo/enterprise#20554
Co-authored-by: fleodoo <fle@odoo.com>