[IMP] documents_spreadsheet: Insert link to odoo menu from spreadsheet
From a spreadsheet, you can now insert a link to any menu in odoo.

Menus are referenced by their xml id. It allows to reference them in spreadsheet
templates (which must be db-independent). Menus that don't have an xml id
(because they were manually created) are referenced with their ids.

Known issue
-----------

The field used to select the ir.menu allows to select a menu you actually
don't have the required access rights to see. That because you have read
access to all "ir.ui.menu" records, even though seeing the menu in the web
UI is restricted to particular user groups.
We can restrict menus with a domain to only match those with the same groups
as the user you might say!
You're right and we actually do that... but there are some additional
restrictions which cannot be handled with a domain. Some menus are blacklisted
with hardcoded behaviors in python. See `_load_menus_blacklist`. We won't
adapt the generic many2one field.
If you select one such menu, the link will display an error telling you that
the menu is unknown.

Task 2572184

Part-of: odoo/enterprise#20554
Co-authored-by: fleodoo <fle@odoo.com>