[IMP] sale_planning,web_gantt: plan slots from gantt view
With this commit, we add a magnifying glass to plan unplanned slots from
the gantt view.

The gantt javascript library calls the `write` method when a user uses
the magnifying glass. To keep this functionality has it is, the method
has to know, in planning slot, whether the write is called on an
unscheduled slot, that is, a slot without start_datetime.
If this unscheduled slot is dependant of a sol with a plannable product,
the write method has to be tuned : slots have to be generated to fill
the remaining hours to plan of the sol.

In web_gantt, the only modification is to call a callback after the
reschedule function in order to display some notification.

The computation of the intervals where slots can be planned take into
account the employee's attendances and leaves, and the slots that are
already planned for him/her.

After this commit, slots will be also planned if there is another slot
with an allocated percentage less than 100 % in the same time, but it
will not generate a conflict (i.e. it will have an allocated percentage
that will fill the gap up-to 100%).

The logic of this algorithm is explained with further explanation in the
code.

The slots generated are also merged at the end of the process in order
to have longer forecasted slots. This doesn't impact the
allocated_hours.

PR community : odoo/odoo#75553
PR enterprise : #19647
PR upgrade : odoo/upgrade#2777

task-2559061 - part:5