[IMP] helpdesk: add sequence for helpdesk ticket
PURPOSE:

In helpdesk, ticket UUID is an increasing numerical value. For Ex: if your ticket
was the nth ticket submitted, the ticket's UUID would be Ticket (#n). For
businesses, this turns out to be a problem for a few reasons. One of the reason
is if user wants to re-format their ticket's UUID they cannot.

SPECIFICATION:

Add a sequence for helpdesk tickect to allow users for the customization of the
number and format of the ticket id.

task-2531499

closes odoo/enterprise#21047

Related: odoo/upgrade#2935
Signed-off-by: Laurent Stukkens (ltu) <ltu@odoo.com>