[IMP] documents_spreadsheet: Insert link to any view in spreadsheet
From any view, you can insert a link to the given view, with the domain,
groups, etc. you just setted. It's actually the same thing as an "ir.filter".

However we have one major constraint: it must work with spreadsheet templates.
It implies that links must be db-independent. We cannot have hardcoded
records ids.

The solution we have for now is to store the values (which are similar to
"ir.filter" values) in the spreadsheet itself.

Known issues
------------

1. One major drawback of the current approach is the migration of this data.
It won't be part of the generic migration process (at least for now).

2. Even though the domain is "saved" in the link and is correctly applied
when clicking on the link to open the view, the domain is not visible in
the search bar. It makes it hard to known what records you are actually seeing.
The context is lost. It's up to the user to annotate the link in the spreadsheet.
This will hopefully be improved in the future. It was considered acceptable
for a first version (FP).

Task 2572184

closes odoo/enterprise#20554

Related: odoo/odoo#75859
Signed-off-by: Vincent Schippefilt (vsc) <vsc@odoo.com>
Co-authored-by: pro-odoo <pro@odoo.com>