[FIX] planning: fix read_group lazy override
Prior to this commit, the read_group lazy argument was forced to False
due to the read_group override.

The read_group override is only useful when the lazy argument is set to
False, the override is now executed when lazy argument has such
value.

read_group params in the super call are clarified to avoid any
ambiguity.

This commit fixes PR #19695 .

task-2633228

closes odoo/enterprise#20473

Signed-off-by: LTU-Odoo <IT-Ideas@users.noreply.github.com>