[IMP] quality, quality_control: quality checks at stock move line level
This creates the possibility to use Quality Checks at the stock move line level.
This can be done in the Control Point, configuring it to track QC at the Lot/SN level.
It is also possible to set a percentage of a lot that has to be checked, indicating
which quantity should be checked within a lot at picking.
The Quality Check pop-up has been transformed into a wizard and adapted.
It shows how many checks need to be done and to which Lot/SN the QC belongs.
It also gives the possibility to go to the previous/next QC.
QC icons have been added in the detailed operations tab, next to the stock move lines having a QC.

closes odoo/enterprise#20102

Task-id: 2566811
Pr: https://github.com/odoo/enterprise/pull/20102
Com-pr: https://github.com/odoo/odoo/pull/75691
Related: odoo/odoo#75691
Related: odoo/upgrade#2785
Signed-off-by: Arnold Moyaux <amoyaux@users.noreply.github.com>