[IMP] documents_spreadsheet: extends global filters to list
Since 2882679a8addbdcea3bfeebf33e166eb75287eb0, it's possible to add
a list view in spreadsheet, but it was not compatible with the global
filters feature.

This commit adapt the global filters to the list. A global filter can filter
values on each pivots and each lists of the spreadsheet.

From a technical point of view, commands relative to global filter has been
renamed from XX_PIVOT_FILTER to XX_GLOBAL_FILTER, the evaluation part (computed domain)
has been moved in data source, and computed in filters_evaluation_plugin.

Task-id 2632018

closes odoo/enterprise#20453

Signed-off-by: Vincent Schippefilt (vsc) <vsc@odoo.com>