[IMP] sale_planning: Avoid computing twice the slots allocated_hours
Purpose
=======

When calling the read_group method, the allocated_hours are computed
for all the slots, to be set with a wrong value.

Then, the value is recomputed again, to be set with the correct value,
obtained from the 'planning_hours_planned' column, which is in the
dependencies of 'allocated_hours'.

By skipping the first call to the compute method, then marking the fields
as to be recomputed for the related records, the is a substential gain
concerning the performances.