[FIX] industry_fsm_{sale,stock}: manage the products quantities
If the SO associated to the task has been modified/confirmed, the
"Choose Products" page won't behave correctly when adding/removing some
products

To reproduce the issue 01:
1. Create a task T
2. Open the associated "Choose Products" page
3. Add a product P
4. Open the generated SO
5. Add a new line (!) for product P
    - The line must be linked with T (`task_id`) thanks to Studio, for
instance
6. Back to the associated "Choose Products" page
    - Note that the quantity of P is now 2 (correct)
7. Set the quantity of P to 0

Error: The quantity becomes 1. When editing the quantity, the module
only considers one SO line (the first one)

To reproduce the issue 02:
1. Create a task T
2. Open the associated "Choose Products" page
3. Add a product P
4. Confirm the SO and associated Delivery
5. Back to the associated "Choose Products" page
6. Add a product P

Error: The quantity is now 3 (should be 2). In the SO, a new line has
been created but its quantity is 2 (should be 1)

OPW-2622853

closes odoo/enterprise#20649

X-original-commit: 6ed4490ff51d21c42347bee016a198ac192a55d2
Signed-off-by: LTU-Odoo <IT-Ideas@users.noreply.github.com>
Signed-off-by: Adrien Widart <adwid@users.noreply.github.com>
Signed-off-by: William André (wan) <wan@odoo.com>