[IMP] stock_barcode: scan product in inventory
In the barcode inventory adjustment, when a product is scanned and its
create a new barcode line, if there is an already existing quant for
this product, we get some additional informations about the quant as its
`inventory_date` and its `quantity` (theoretical qty. on hand).

task-2632818

X-original-commit: d1c939d071a437f10e4df5500d33de9fb93af747
Part-of: odoo/enterprise#21810