[IMP] project_forecast, _*: improve the planning forecast stat button
_* = industry_fsm, industry_fsm_report, project_enterprise

Currently, the current week is opened in the Gantt of planning from the
forecast stat button of the project.

Planning gantt view should open on the week the start date falls,
if it is in the future else, it should open on the current week.

Removed the display_fsm_dates field on the project task.

The string date of the project date field has been changed to
the planned date.

closes #19760

Taskid: 2532909
Related: odoo/upgrade#2665
Signed-off-by: Yannick Tivisse (yti) <yti@odoo.com>