[FIX] account_reports: set previous period on custom date
Steps to reproduce :

Go to Accounting -> Reporting -> Balance Sheet
Select any date in the past
Select Comparison -> Previous Period
Issue :

The previous periods displayed are not related with the choosen date,
but with the actual date.

Thanks to @smetl and some refactoring, we are now able to get the right
previous_periods when setting a custom date.

opw-2616662

closes odoo/enterprise#20873

X-original-commit: 2e88684174659c22da5aac5c40cd33c204f02023
Signed-off-by: Laurent Smet <smetl@users.noreply.github.com>