[FIX] web_studio: No traceback when click in a field contained in a t-else
The field can be an `owl` (line 143) or a old odoo widget (line 153).
When in studio the `appendTo` method is called, it does not exist and
there is a traceback.
The choice of this fix is to add the `appendTo` method on the owl
components, so as not to have to add a condition depending on the
instance of `owl` (like line 1431) each time.
This change can be removed when web_studio is converted to owl.

opw-2529150
opw-2631717

closes odoo/enterprise#21078

X-original-commit: 84bf0816bc9bd531b00b49f9a234067d753242cb
Signed-off-by: Vincent Schippefilt (vsc) <vsc@odoo.com>