[IMP] partner_commission: Misc improvements
Minimum amount
--------------

Add a new restriction for automatic confirmation of the PO:
The Po should have a minimum amount_total define per company
if a po does not have reach the minimum amount, it will not be confirmed
automatically. So, the po wait for more commission.

This avoid to pay to payment fee for small commission.

Archive
-------
Allow commission plan to be archived

Commission plan multi company
-----------------------------
Commission plan can be link to only one company,
in that case it won't be possible to assign the plan
to a subscription or sale order from another company.

+ Fix the settings

closes odoo/enterprise#20535

Related: odoo/upgrade#2790
Signed-off-by: Arnaud Joset <arj-odoo@users.noreply.github.com>