[ADD] sale_ebay_accont_deletion: support ebay account deletion requests
From September 1, 2021, all eBay integration must have an endpoint to
be notified when an eBay Marketplace User requests to delete his data.

This stable module provides this endpoint, and the necessary setup
instructions.

When a account deletion notification is received, linked partner(s) are
anonymized, and an activity is scheduled for the admin to make sure
the remaining data is removed or can be kept according to the laws.

task-2504306

closes odoo/enterprise#20652

X-original-commit: 0bb7f158a4c8348e6687edc9aff2d120ec7f9b72
Signed-off-by: Victor Feyens (vfe) <vfe@odoo.com>
Co-authored-by: Valentin Chevalier <vcr@odoo.com>