[IMP] documents_spreadsheet: Only join values when grouping by date
currently when adding a pivot to a spreadsheet, each subgroup has always the same set of values.
this join of all possible values in a set has been most valuable when first grouping by a date field,
but can be quite frustrating when grouping by fields that are not independent, but related.

This commit changes the behavior off pivots in order to only have the set of subgroups
to be the union of all possible values if there is a parent group based on a date field.

There is also a special case if both parent and child are based on the same date field.
in that cas we wont join the values.

closes odoo/enterprise#18933

Task-id: 251015
Signed-off-by: Pierre Rousseau (pro) <pro@odoo.com>