[FIX] sign: fixes resize causing signature in edit mode In task-2475732, a feature was added that allowed users to edit sign requests while signing. This consists of adding sign items, moving them with drag and drop, resizing them and signing them after they are correctly placed. After signing a signature/initial field, it is not possible to move it or resize it.
Due to the differences in the HTML content between the editable sign item and
the non-editable sign item, the click event registered to sign each item should
be different. Currently, the click event is registered on the sign item itself,
for both cases. However resizing causes the event to bubble up and trigger the
click event in the sign item (afterall resize and drag-drop are just mousedown
and mouseup events). This then causes the signature/opening of wizard everytime
a resize or a drag is done in the added sign items.

This PR solves this problem with the following approach: register the click
event in the "o_sign_config_area" element instead of the sign item, when the
sign item is editable.

task-2636460

closes odoo/enterprise#20574

Signed-off-by: Arnaud Joset <arj-odoo@users.noreply.github.com>