[FIX] stock_barcode: can scroll on form view
Before this commit, in case the stock barcode move line form view is too
long (it can happen if the quants list contains too much quants for
example), it was not possible to scroll.
The issue only occurred in the desktop mode, as the scroll was already
enabled for the mobile environment.

Part-of: odoo/enterprise#15755