[FIX] account_reports: tax report: fix wrong use of generic ids
The 'Sales' and 'Purchase' lines of the generic tax report were built as if their related model was an account.tax ; it was wrong: those lines are nothing but a markup, they are not created from any model. The model name stored in the line id should be None.

Part-of: odoo/enterprise#19628