[FIX] account_batch_payment: UX clean payments popup
Set journal, payment method, batch payment fields to optional hide in batch payment many2many widget popup view (simplify the UI)

The manifest is reordered so that the payment views are loaded before the batch payment views. This is because batch payment should be visible by default in the payments tree view, but optional hide in the payments tree view that shows up when adding a line to the batch payments. In order to do this, batch payment id field is loaded into the payment view first using inheritance, and then in batch payment views an alternative payment tree view is also provided, using primary inheritance. This optional hides the aformentioned batch_payment_id, as well as the journal_id and the payment_method_line_id, such that they are optionally hidden only in the popup view when adding lines to the batch payment.

Task: 2613517
Closes: #20581
Signed-off-by: Quentin De Paoli <qdp@odoo.com>